backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "llm-backend"
version = "0.1.0"
description = "Unified LLM Backend"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn",
    "pydantic",
    "vertexai",
    "google-cloud-bigquery",
    "google-cloud-storage",
    "psycopg2-binary",
    "pgvector",
    "beautifulsoup4",
    "requests",
    "pymupdf",
    "pandas",
    "python-dotenv",
]

[tool.hatch.build.targets.wheel]
packages = ["."]
